Call repair only if app is updated
